resources/validate-authentication-aws-cognito/requirements.txt
ecdsa==0.14.1
pyasn1==0.4.8
python-jose==3.2.0
rsa==4.7.2
six==1.15.0
